Fallback to luaposix as backup
In some circumstantces it is desireable to depend on pre-installed
lua-posix support instead of building the internal dependency. This
patch allows the module to fall back on luaposix's isatty()
implementation instead of using the simple native wrapper. This
implementation is slightly slower than the builtin native wrapper, but
using it means that it is possible to load this as a pure lua module
if luaposix is already installed.